requirements.txt
Flask>=2.2
Flask-WTF>=1.0
PyYAML>=6.0
scapy>=2.4.5
WTForms>=3.0